# common-platform-product-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-product-mgmt
  description: Product Management Core Application
  contact:
    name: Catalis Development
    email: dev@catalis.com
  version: 1.0.0
servers:
  - url: 'http://core.catalis.vc/products'
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  '/api/v1/fee-structures/{feeStructureId}/components/{componentId}':
    get:
      tags:
        - Product Fee Component
      summary: Get Fee Component by ID
      description: Retrieve a specific fee component by its unique identifier.
      operationId: getFeeComponent
      parameters:
        - name: feeStructureId
          in: path
          description: Unique identifier of the fee structure
          required: true
          schema:
            type: integer
            format: int64
        - name: componentId
          in: path
          description: Unique identifier of the fee component
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved the fee component
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeeComponentDTO'
        '404':
          description: Fee component not found
    put:
      tags:
        - Product Fee Component
      summary: Update Fee Component
      description: Update an existing fee component by its unique identifier.
      operationId: updateFeeComponent
      parameters:
        - name: feeStructureId
          in: path
          description: Unique identifier of the fee structure
          required: true
          schema:
            type: integer
            format: int64
        - name: componentId
          in: path
          description: Unique identifier of the fee component to update
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeeComponentDTO'
        required: true
      responses:
        '200':
          description: Fee component updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeeComponentDTO'
        '404':
          description: Fee component not found
    delete:
      tags:
        - Product Fee Component
      summary: Delete Fee Component
      description: Remove an existing fee component by its unique identifier.
      operationId: deleteFeeComponent
      parameters:
        - name: feeStructureId
          in: path
          description: Unique identifier of the fee structure
          required: true
          schema:
            type: integer
            format: int64
        - name: componentId
          in: path
          description: Unique identifier of the fee component to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Fee component deleted successfully
        '404':
          description: Fee component not found
...
  '/api/v1/fee-structures/{feeStructureId}/components/filter':
    post:
      tags:
        - Product Fee Component
      summary: List Fee Components by Fee Structure
      description: >-
        Retrieve a paginated list of fee components associated with a specific
        fee structure.
      operationId: getByFeeStructureId
      parameters:
        - name: feeStructureId
          in: path
          description: Unique identifier of the fee structure
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaginationRequest'
        required: true
      responses:
        '200':
          description: Successfully retrieved fee components
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseFeeComponentDTO'
        '404':
          description: No fee components found for the specified structure
  '/api/v1/fee-structures/{feeStructureId}/components':
    post:
      tags:
        - Product Fee Component
      summary: Create Fee Component
      description: 'Create a new fee component, which can be attached to a fee structure.'
      operationId: createFeeComponent
      parameters:
        - name: feeStructureId
          in: path
          description: Unique identifier of the fee structure
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeeComponentDTO'
        required: true
      responses:
        '200':
          description: Fee component created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeeComponentDTO'
...
components:
  schemas:
...
    FeeComponentDTO:
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        feeComponentId:
          type: integer
          format: int64
          readOnly: true
        feeStructureId:
          type: integer
          format: int64
        feeType:
          type: string
          enum:
            - ORIGINATION_FEE
            - LATE_PAYMENT_FEE
            - TRANSACTION_FEE
            - SERVICE_FEE
        feeDescription:
          type: string
        feeAmount:
          type: number
        feeUnit:
          type: string
          enum:
            - FIXED
            - PERCENTAGE
        applicableConditions:
          type: string
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: 'The direction of sorting, either ASC or DESC.'
          example: DESC
          default: DESC
      description: >-
        Represents a pagination request for retrieving paginated results,
        including page number, size, sort field, and direction.
...
    PaginationResponseFeeComponentDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/FeeComponentDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: 'The current page number, typically zero-based.'
          format: int32
      description: >-
        Represents a paginated response containing a list of items and
        pagination metadata.
...